+++ packages/analytics-js-integrations/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.10.0](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.9.3...@rudderstack/analytics-js-integrations@3.10.0) (2024-10-25)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.11.0`
+
+### Features
+
+* add session replay for mixpanel ([#1899](https://github.com/rudderlabs/rudder-sdk-js/issues/1899)) ([bfa9d0a](https://github.com/rudderlabs/rudder-sdk-js/commit/bfa9d0a2c8e9e821b32da0defa5c779e090bd106))
+* gainsight PX destination ([#1852](https://github.com/rudderlabs/rudder-sdk-js/issues/1852)) ([#1889](https://github.com/rudderlabs/rudder-sdk-js/issues/1889)) ([3a705f0](https://github.com/rudderlabs/rudder-sdk-js/commit/3a705f063bcae99c7964495ff83ad9ce8d4eb5a3))
+
 ## [3.9.3](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-integrations@3.9.2...@rudderstack/analytics-js-integrations@3.9.3) (2024-10-21)
 
 ### Dependency Updates